+++ inventory-service/src/main/java/ma/xproce/inventoryservice/dao/entities/Video.java
@@ -15,7 +15,7 @@ import java.util.Date;
 @Builder
 public class Video {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Float id;
     private String name;
     private String url;
     private String description;
