+++ inventory-service/src/main/java/ma/xproce/inventoryservice/dao/repositories/CreatorRepository.java
@@ -4,5 +4,5 @@ import ma.xproce.inventoryservice.dao.entities.Creator;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CreatorRepository extends JpaRepository<Creator,Long> {
-    Creator findCreatorById(Long id);
+    Creator findCreatorById(Float id);
 }
